requirements.txt
azure-functions
requests
python-docx
openpyxl
PyPDF2
